+++ batch/src/test/java/kr/co/techcourse/prolog/batch/configuration/schedule/fixture/TestConfiguration_2.java
@@ -1,6 +1,6 @@
 package kr.co.techcourse.prolog.batch.configuration.schedule.fixture;
 
-import kr.co.techcourse.prolog.batch.job.common.EnableBatchJob;
+import kr.co.techcourse.prolog.batch.configuration.schedule.EnableBatchJob;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.core.JobParametersIncrementer;
